requirements.txt
paramiko~=3.2.0
requests~=2.31.0
mysql-connector~=2.2.9
mysql-connector-python~=8.0.33
PyMySQL~=1.0.3
google-api-python-client~=2.88.0
google-auth-oauthlib~=1.0.0
pandas~=2.0.2
selenium~=4.10.0
SQLAlchemy~=2.0.15
python-telegram-bot~=20.3
google-auth~=2.19.1